Heavy refactoring
Cleaned up all the files

The code is easier to read. The functions have become smaller and
properly grouped up in classes. I am really happy with this iteration of
the code. It doesn't really add much more functionality, but the code
looks so much cleaner. There are still a few todos. Specially the
image_thread.py file needs to be looked at.

- Added config file instead of hardcoded settings